Intro new command 'change registry' to re-configure registry and credentials wso2/k8s-apim-operator#184
