+++ import-export-cli/cmd/installApiOperator.go
@@ -19,13 +19,7 @@
 package cmd
 
 import (
-	"encoding/base64"
-	"encoding/json"
 	"fmt"
-	"github.com/cbroglie/mustache"
-	"github.com/wso2/product-apim-tooling/import-export-cli/box"
-	"gopkg.in/yaml.v2"
-	"strings"
 	"time"
 
 	"github.com/spf13/cobra"
@@ -54,17 +48,19 @@ to quickly create a Cobra application.`,
 	Run: func(cmd *cobra.Command, args []string) {
 		utils.Logln(utils.LogPrefixInfo + installApiOperatorCmdLiteral + " called")
 
-		// OLM installation requires time to install before installing the WSO2 API Operator
-		// Hence getting user inputs
-		registryUrl, repository, username, password := readInputs()
+		// read inputs for docker registry
+		registryUrl, username, password := readDockerRegistryInputs()
+
 		isLocalInstallation := flagApiOperatorFile != ""
 		if !isLocalInstallation {
 			installOLM(utils.OlmVersion)
 			installApiOperatorOperatorHub()
 		}
 
+		createControllerConfigs(isLocalInstallation)
+		setRegistryRepositoryOnControllerConfig(registryUrl)
 		createDockerSecret(registryUrl, username, password)
-		createControllerConfigs(repository, isLocalInstallation)
+		setToK8sMode()
 	},
 }
 
@@ -134,40 +130,8 @@ func installApiOperatorOperatorHub() {
 	}
 }
 
-// createDockerSecret creates K8S secret with credentials for docker registry
-func createDockerSecret(registryUrl string, username string, password string) {
-	encodedCredential := base64.StdEncoding.EncodeToString([]byte(username + ":" + password))
-	auth := map[string]map[string]map[string]string{
-		"auths": {
-			registryUrl: {
-				"auth":     encodedCredential,
-				"username": username,
-				"password": password,
-			},
-		},
-	}
-	authJsonByte, err := json.Marshal(auth)
-	if err != nil {
-		utils.HandleErrorAndExit("Error marshalling docker secret credentials ", err)
-	}
-
-	encodedAuthJson := base64.StdEncoding.EncodeToString(authJsonByte)
-	secretTemplate, _ := box.Get("/kubernetes_resources/registry_secret_mustache.yaml")
-	secretYaml, err := mustache.Render(string(secretTemplate), map[string]string{
-		"encodedJson": encodedAuthJson,
-	})
-	if err != nil {
-		utils.HandleErrorAndExit("Error rendering docker secret credentials", err)
-	}
-
-	// apply created secret yaml file
-	if err := utils.K8sApplyFromStdin(secretYaml); err != nil {
-		utils.HandleErrorAndExit("Error creating docker secret credentials", err)
-	}
-}
-
-// createControllerConfigs downloads the mustache, replaces repository value and creates the config: `controller-config`
-func createControllerConfigs(repository string, isLocalInstallation bool) {
+// createControllerConfigs creates configs
+func createControllerConfigs(isLocalInstallation bool) {
 	utils.Logln(utils.LogPrefixInfo + "Installing controller configs")
 	fmt.Println("Installing controller configurations...")
 	configFile := flagApiOperatorFile
@@ -189,103 +153,14 @@ func createControllerConfigs(repository string, isLocalInstallation bool) {
 			utils.HandleErrorAndExit("Error creating configurations", err)
 		}
 	}
-
-	// get controller config config map
-	controllerConfigMapYaml, err := utils.GetCommandOutput(
-		utils.Kubectl, utils.K8sGet, "cm", utils.ApiOpControllerConfigMap,
-		"-n", utils.ApiOpWso2Namespace,
-		"-o", "yaml",
-	)
-	if err != nil {
-		utils.HandleErrorAndExit("Error reading controller-config", err)
-	}
-
-	// replace registry
-	controllerConfigMap := make(map[interface{}]interface{})
-	if err := yaml.Unmarshal([]byte(controllerConfigMapYaml), &controllerConfigMap); err != nil {
-		utils.HandleErrorAndExit("Error reading controller-config", err)
-	}
-	controllerConfigMap["data"].(map[interface{}]interface{})["dockerRegistry"] = repository
-	configuredConfigMap, err := yaml.Marshal(controllerConfigMap)
-	if err != nil {
-		utils.HandleErrorAndExit("Error rendering controller-config", err)
-	}
-
-	// apply controller config config map back
-	if err := utils.K8sApplyFromStdin(string(configuredConfigMap)); err != nil {
-		utils.HandleErrorAndExit("Error creating controller-configs", err)
-	}
 }
 
-// readInputs reads docker-registry URL, repository, username and password from the user
-func readInputs() (string, string, string, string) {
-	isConfirm := false
-	registryUrl := ""
-	repository := ""
-	username := ""
-	password := ""
-	var err error
-
-	for !isConfirm {
-		registryUrl, err = utils.ReadInputString("Enter Docker-Registry URL", utils.DockerRegistryUrl, "", true)
-		if err != nil {
-			utils.HandleErrorAndExit("Error reading Docker-Registry URL", err)
-		}
-
-		repository, err = utils.ReadInputString("Enter Repository Name", "", utils.UsernameValidationRegex, true)
-		if err != nil {
-			utils.HandleErrorAndExit("Error reading Repository Name", err)
-		}
-
-		username, err = utils.ReadInputString("Enter Username", "", utils.UsernameValidationRegex, true)
-		if err != nil {
-			utils.HandleErrorAndExit("Error reading Username", err)
-		}
-
-		password, err = utils.ReadPassword("Enter Password")
-		if err != nil {
-			utils.HandleErrorAndExit("Error reading Password", err)
-		}
-
-		isCredentialsValid, err := validateDockerRegistry(registryUrl, repository, username, password)
-		if !isCredentialsValid {
-			utils.HandleErrorAndExit("Error connecting to Docker Registry repository using credentials", err)
-		}
-
-		fmt.Println("")
-		fmt.Println("Docker-Registry URL: " + registryUrl)
-		fmt.Println("Repository         : " + repository)
-		fmt.Println("Username           : " + username)
-
-		isConfirmStr, err := utils.ReadInputString("Confirm configurations", "Y", "", false)
-		if err != nil {
-			utils.HandleErrorAndExit("Error reading user input Confirmation", err)
-		}
-
-		isConfirmStr = strings.ToUpper(isConfirmStr)
-		isConfirm = isConfirmStr == "Y" || isConfirmStr == "YES"
-	}
-
-	return registryUrl, repository, username, password
-}
-
-// validateDockerRegistry validates the credentials against registry url and repository
-func validateDockerRegistry(registryUrl string, repository string, username string, password string) (bool, error) {
-	// remove version tag if exists
-	//regExpString := `\/v[\d.-]*\/?$`
-	//regExp := regexp.MustCompile(regExpString)
-	//registryUrl = regExp.ReplaceAllString(registryUrl, "")
-	//
-	//hub, err := registry.New(registryUrl, username, password)
-	//if err != nil {
-	//	return false, err
-	//}
-	//
-	//if _, err := hub.Repositories(); err != nil {
-	//	return false, err
-	//}
-
-	return true, nil
+// setToK8sMode sets the apictl mode to kubernetes
+func setToK8sMode() {
+	// read the existing config vars
+	configVars := utils.GetMainConfigFromFile(utils.MainConfigFilePath)
+	configVars.Config.KubernetesMode = true
+	utils.WriteConfigFile(configVars, utils.MainConfigFilePath)
 }
 
 // init using Cobra
